\name{sgr-package}
\alias{sgr-package}
\alias{sgr}
\docType{package}
\title{
Sample Generation by Replacement
}
\description{
The library for Sample Generation by Replacement simulations. The package can be used to perform fake data analysis according to the sample generation by replacement approach. It includes functions for making simple inferences about
discrete/ordinal fake data. The package allows to study the implications of fake data for empirical results.
}
\details{
\tabular{ll}{
Package: \tab sgr\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Suggests: \tab MASS, polycor\cr
Date: \tab 2013-06-09\cr
License: \tab GPL (>=2)\cr
Encoding: \tab UTF-8\cr
}
}
\author{
Massimiliano Pastore & Luigi Lombardi

Maintainer: Massimiliano Pastore <massimiliano.pastore@unipd.it>
}
\references{
Lombardi, L., Pastore, M. (2012). Sensitivity of fit indices to fake perturbation of ordinal data: A sample by replacement approach. \emph{Multivariate Behavioral Research}, 47, 519-546.

Pastore, M., Lombardi, L. (in press). The impact of faking on Cronbach’s Alpha for
dichotomous and ordered rating scores. \emph{Quality & Quantity}. 
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
