% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGR_undirected.R
\name{SGR_undirected}
\alias{SGR_undirected}
\title{Spatial Graphical Regression function (undirected setting)}
\usage{
SGR_undirected(features, spatial_locations, ncores, N_MCMC, N_BURNIN, N_thin)
}
\arguments{
\item{features}{This is gene expression matrix of dimension nXp}

\item{ncores}{Number of cores}

\item{N_MCMC}{Number of MCMC iterations to be saved}

\item{N_BURNIN}{Number of burnin iterations (before start to save)}

\item{N_thin}{Number of iterations to skip between two saved iterations}

\item{Spatial_locations}{Spatial coordinates nX2}
}
\value{
A list containing mcmc samples of for a node specific regression.
(1) Beta_post: Posterior samples of beta
(2) gamma_prob_post: Posterior samples of probability of selection indicators
(3) sigma_epsilon_post: Posterior samples of error variance
}
\description{
Spatial Graphical Regression for undirected setting
     This code is based on MCMC algorithm
}
