% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimBaS.R
\name{SimBaS}
\alias{SimBaS}
\title{SimBaS}
\usage{
SimBaS(f_mcmc_samples, spatial_locations)
}
\arguments{
\item{f_mcmc_sample}{MCMC samples of the function f (n X n_mcmc)}

\item{Spatial_locations}{Spatial coordinates nX2}
}
\value{
a list containing:
        P_SimBaS <- local edge probability for each spatial locations.
        f_non_zero <- functional values where it is bounded away from zero.
        significant_locations <- locations where the function is away from zero.
}
\description{
Implementation of SimBaS steps to find out which
part of the functions are bounded away from zero.
}
