% --- Source file: smokers.Rd --- 
\name{smokers} 
\alias{smokers} 
\title{Data set} 

\description{ 
Data was collected
using a single item selected from a survey regarding the use of alcohol and
other drugs in adolescents.} 

\usage{data(smokers)} 

\format{
This data frame contains the following columns:
    \itemize{
    \item \code{age}: int, 1 = adults, 2 = minors.
    \item \code{smoking}: int, 1 = no, 2 = yes. 
    \item \code{drug}: int, drug addiction, 1 = no, 2 = yes. 
}
}
\source{Pastore, M., Lombardi, L., Mereu, F. (2007). Effects of malingering in self-report measures: A scenario analysis approach; in C. H. Skiadas (Ed.), \emph{Recent Advances in Stochastic Modeling and Data Analysis}, pp. 483-491, World Scientific Publishing.}
%\author{Massimiliano Pastore}
\keyword{data}
