% --- Source file: smokers.Rd --- 
\name{smokers} 
\alias{smokers} 
\title{Data set} 

\description{ 
Data was collected using a single item selected from a survey about the use of alcohol and other drugs among young people.} 

\usage{data(smokers)} 

\format{
This data frame contains the following columns:
    \itemize{
    \item \code{age}: int, 1 = adults, 2 = minors.
    \item \code{smoking}: int, 1 = yes, 2 = no. 
    \item \code{drug}: int, drug addiction, 1 = yes, 2 = no. 
    \item \code{druguse}: int, drug consumption, 1 = never, 2 = once, 3 = some times, 
        4 = often.
}
}
\source{Pastore, M., Lombardi, L., Mereu, F. (2007). Effects of malingering in self-report measures: A scenario analysis approach; in C. H. Skiadas (Ed.), \emph{Recent Advances in Stochastic Modeling and Data Analysis}, pp. 483-491, World Scientific Publishing.}
%\author{Massimiliano Pastore}
\keyword{data}
